Refactored admin and agent authentication routes: moved signup and login endpoints to auth router
- Moved agent signup and login from agentRoutes.js to authRoutes.js under '/auth/agents/signup' and '/auth/agents/login'.
- Moved admin signup and login from adminRoutes.js to authRoutes.js under '/auth/admins/signup' and '/auth/admins/login'.
- Ensured consistency and separation of concern between user-specific routes (agents, admins) and authentication functionality.
- Updated route structure for clarity and better API management.
