+++ server/routes/authRoutes.js
@@ -21,6 +21,22 @@ import { validateSignUpFields } from '../middleware/validation/signupValidation.
 
 import authenticateJWT from '../middleware/auth/authMiddleware.js';
 
+import {
+  createAdmin,
+  loginAdmin,
+} from '../controllers/auth/adminAuthController.js';
+import { validateAdminFields } from '../middleware/validation/adminValidation.js';
+import {checkIsSuperAdmin} from '../middleware/auth/roleMiddleware.js';
+
+import {
+  createAgent,
+  loginAgent,
+} from '../controllers/auth/agentAuthController.js';
+import { handleValidationErrors } from '../middleware/common/handleValidationErrors.js';
+import { validateAgentFields } from '../middleware/validation/agentValidation.js';  // Input validation for agents
+
+
+
 const router = express.Router();
 
 /**
@@ -30,10 +46,15 @@ const router = express.Router();
  */
 
 // Google OAuth Sign-Up Route
-router.get('/google/signup', passport.authenticate('google-signup'));
+router.get(
+  '/google/signup', 
+  passport.authenticate('google-signup'));
 
 // Google OAuth Sign-In Route
-router.get('/google/signin', passport.authenticate('google-signin'));
+router.get(
+  '/google/signin', 
+  passport.authenticate('google-signin')
+);
 
 // Google OAuth Callback for Sign Up
 router.get(
@@ -95,5 +116,47 @@ router.post('/reset-password', resetPassword);
 // Verify if the JWT token is still valid
 router.get('/verify-token', authenticateJWT, verifyToken);
 */
+
+/**
+ * ============================
+ * Admin Authentication Routes
+ * ============================
+ */
+
+// Create a new admin (super admin only)
+router.post(
+  '/admins/signup',
+  authenticateJWT, 
+  checkIsSuperAdmin,
+  validateAdminFields,
+  handleValidationErrors,  
+  createAdmin
+);
+
+// Admin Login
+router.post(
+  '/admins/login',
+   loginAdmin
+);
+
+/**
+ * ============================
+ * Agent Authentication Routes
+ * ============================
+ */
+
+// Register a new agent (public route)
+router.post(
+  '/agents/signup',
+  validateAgentFields,
+  handleValidationErrors,  
+  createAgent
+);
+
+// Agent login route (public)
+router.post(
+  '/agents/login',
+  loginAgent
+);
 
 export default router;
